refactore: add scripts folder
