+++ package.json
@@ -24,6 +24,8 @@
   "devDependencies": {
     "eslint": "^8.24.0",
     "eslint-config-prettier": "^8.5.0",
+    "postcss": "^8.4.17",
+    "postcss-scss": "^4.0.5",
     "sass": "^1.55.0",
     "vite": "^3.1.0",
     "vite-plugin-html": "^3.2.0"
