--- src/pages/dent/dent.scss
@@ -1,1 +1,0 @@
-@use '/style/abstract' as *;
